calcGseaStatBatchCpp {fgsea}	R Documentation

Calculates GSEA statistic valus for all gene sets in 'selectedStats' list.

Description

Takes O(n + mKlogK) time, where n is the number of genes, m is the number of gene sets, and k is the mean gene set size.

Usage

calcGseaStatBatchCpp(stats, selectedGenes, geneRanks)

Arguments

stats
Numeric vector of gene-level statistics sorted in decreasing order

selectedGenes
List of integer vector with integer gene IDs (from 1 to n)

geneRanks
Integer vector of gene ranks

Value

Numeric vector of GSEA statistics of the same length as 'selectedGenes' list

[Package fgsea version 1.20.0 Index]