fgseaLabel {fgsea}	R Documentation

Runs label-permuring gene set enrichment analysis.

Description

Runs label-permuring gene set enrichment analysis.

Usage

fgseaLabel(
  pathways,
  mat,
  labels,
  nperm,
  minSize = 1,
  maxSize = Inf,
  nproc = 0,
  gseaParam = 1,
  BPPARAM = NULL
)

Arguments

pathways
List of gene sets to check.

mat
Gene expression matrix. Row name should be the same as in 'pathways'

labels
Numeric vector of labels for the correlation score of the same length as the number of columns in 'mat'

nperm
Number of permutations to do. Minimial possible nominal p-value is about 1/nperm

minSize
Minimal size of a gene set to test. All pathways below the threshold are excluded.

maxSize
Maximal size of a gene set to test. All pathways above the threshold are excluded.

nproc
If not equal to zero sets BPPARAM to use nproc workers (default = 0).

gseaParam
GSEA parameter value, all gene-level statis are raised to the power of 'gseaParam' before calculation of GSEA enrichment scores.

BPPARAM
Parallelization parameter used in bplapply. Can be used to specify cluster to run. If not initialized explicitly or by setting 'nproc' default value 'bpparam()' is used.

Value

A table with GSEA results. Each row corresponds to a tested pathway. The columns are the following:

pathway – name of the pathway as in 'names(pathway)';

pval – an enrichment p-value;

padj – a BH-adjusted p-value;

ES – enrichment score, same as in Broad GSEA implementation;

NES – enrichment score normalized to mean enrichment of random samples of the same size;

nMoreExtreme' – a number of times a random gene set had a more extreme enrichment score value;

size – size of the pathway after removing genes not present in 'names(stats)'.

leadingEdge – vector with indexes of leading edge genes that drive the enrichment, see http://software.broadinstitute.org/gsea/doc/GSEAUserGuideTEXT.htm#_Running_a_Leading.

Examples

library(limma)
library(GEOquery)
es <- getGEO("GSE19429", AnnotGPL = TRUE)[[1]]
exprs(es) <- normalizeBetweenArrays(log2(exprs(es)+1), method="quantile")
es <- es[!grepl("///", fData(es)$`Gene ID`), ]
es <- es[fData(es)$`Gene ID` != "", ]
es <- es[order(apply(exprs(es), 1, mean), decreasing=TRUE), ]
es <- es[!duplicated(fData(es)$`Gene ID`), ]
rownames(es) <- fData(es)$`Gene ID`

pathways <- reactomePathways(rownames(es))
mat <- exprs(es)
labels <- as.numeric(as.factor(gsub(" .*", "", es$title)))
fgseaRes <- fgseaLabel(pathways, mat, labels, nperm = 1000, minSize = 15, maxSize = 500)

[Package fgsea version 1.20.0 Index]