fgseaSimpleImpl {fgsea}	R Documentation

Runs preranked gene set enrichment analysis for preprocessed input data.

Description

Runs preranked gene set enrichment analysis for preprocessed input data.

Usage

fgseaSimpleImpl(
  pathwayScores,
  pathwaysSizes,
  pathwaysFiltered,
  leadingEdges,
  permPerProc,
  seeds,
  toKeepLength,
  stats,
  BPPARAM,
  scoreType
)

Arguments

pathwayScores
Vector with enrichment scores for the 'pathways'.

pathwaysSizes
Vector of pathways sizes.

pathwaysFiltered
Filtered pathways.

leadingEdges
Leading edge genes.

permPerProc
Parallelization parameter for permutations.

seeds
Seed vector

toKeepLength
Number of 'pathways' that meet the condition for 'minSize' and 'maxSize'.

stats
Named vector of gene-level stats. Names should be the same as in 'pathways'

BPPARAM
Parallelization parameter used in bplapply.

scoreType
This parameter defines the GSEA score type. Possible options are ("std", "pos", "neg") Can be used to specify cluster to run. If not initialized explicitly or by setting 'nproc' default value 'bpparam()' is used.

Value

A table with GSEA results. Each row corresponds to a tested pathway. The columns are the following:

pathway – name of the pathway as in 'names(pathway)';

pval – an enrichment p-value;

padj – a BH-adjusted p-value;

ES – enrichment score, same as in Broad GSEA implementation;

NES – enrichment score normalized to mean enrichment of random samples of the same size;

nMoreExtreme' – a number of times a random gene set had a more extreme enrichment score value;

size – size of the pathway after removing genes not present in 'names(stats)'.

leadingEdge – vector with indexes of leading edge genes that drive the enrichment, see http://software.broadinstitute.org/gsea/doc/GSEAUserGuideTEXT.htm#_Running_a_Leading.

[Package fgsea version 1.20.0 Index]