mapIdsList {fgsea}	R Documentation

Effeciently converts collection of pathways using AnnotationDbi::mapIds function. Parameters are the sames as for mapIds except for keys, which is assumed to be a list of vectors.

Description

Effeciently converts collection of pathways using AnnotationDbi::mapIds function. Parameters are the sames as for mapIds except for keys, which is assumed to be a list of vectors.

Usage

mapIdsList(x, keys, column, keytype, ...)

Arguments

x
the AnnotationDb object. But in practice this will mean an object derived from an AnnotationDb object such as a OrgDb or ChipDb object.

keys
a list of vectors with gene ids

column
the column to search on

keytype
the keytype that matches the keys used

...
other parameters passed to AnnotationDbi::mapIds

See Also

AnnotationDbi::mapIds

Examples

library(org.Mm.eg.db)
data(exampleRanks)
fgseaRes <- fgsea(examplePathways, exampleRanks, maxSize=500, eps=1e-4)
fgseaRes[, leadingEdge := mapIdsList(org.Mm.eg.db, keys=leadingEdge, column="SYMBOL", keytype="ENTREZID")]

[Package fgsea version 1.20.0 Index]