multilevelImpl {fgsea}	R Documentation

Calculates P-values for preprocessed data.

Description

Calculates P-values for preprocessed data.

Usage

multilevelImpl(
  multilevelPathwaysList,
  stats,
  sampleSize,
  seed,
  eps,
  sign = FALSE,
  BPPARAM = NULL
)

Arguments

multilevelPathwaysList
List of pathways for which P-values will be calculated.

stats
Named vector of gene-level stats. Names should be the same as in 'pathways'

sampleSize
The size of a random set of genes which in turn has size = pathwaySize

seed
'seed' parameter from 'fgseaMultilevel'

eps
This parameter sets the boundary for calculating the p value.

sign
This option will be used in future implementations.

BPPARAM
Parallelization parameter used in bplapply. Can be used to specify cluster to run. If not initialized explicitly or by setting 'nproc' default value 'bpparam()' is used.

Value

List of P-values.

[Package fgsea version 1.20.0 Index]